refactor: entry point에서는 호출만 수행하고 controller가 view, game 직접 의존
